Break up TestingIntegrationTest and move tests to function-specific test classes
