+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/AbstractJUnitTestExecutionIntegrationTest.groovy
@@ -17,18 +17,22 @@
 package org.gradle.testing.junit
 
 import org.gradle.api.JavaVersion
+import org.gradle.integtests.fixtures.AvailableJavaHomes
 import org.gradle.integtests.fixtures.DefaultTestExecutionResult
 import org.gradle.integtests.fixtures.TestClassExecutionResult
 import org.gradle.integtests.fixtures.TestExecutionResult
+import org.gradle.integtests.fixtures.jvm.JavaToolchainFixture
 import org.gradle.test.fixtures.file.TestFile
 import org.gradle.testing.fixture.AbstractTestingMultiVersionIntegrationTest
+import org.hamcrest.CoreMatchers
+import spock.lang.Issue
 
 import static org.hamcrest.CoreMatchers.containsString
 import static org.hamcrest.CoreMatchers.equalTo
 import static org.hamcrest.CoreMatchers.not
 import static org.hamcrest.MatcherAssert.assertThat
 
-abstract class AbstractJUnitTestExecutionIntegrationTest extends AbstractTestingMultiVersionIntegrationTest {
+abstract class AbstractJUnitTestExecutionIntegrationTest extends AbstractTestingMultiVersionIntegrationTest implements JavaToolchainFixture {
     abstract String getJUnitVersionAssertion()
     abstract TestClassExecutionResult assertFailedToExecute(TestExecutionResult testResult, String testClassName)
 
@@ -303,4 +307,201 @@ abstract class AbstractJUnitTestExecutionIntegrationTest extends AbstractTesting
         DefaultTestExecutionResult testResult = new DefaultTestExecutionResult(testDirectory)
         assertFailedToExecute(testResult, 'com.example.Foo').assertTestCount(1, 1, 0)
     }
+
+    @Issue("https://issues.gradle.org/browse/GRADLE-1948")
+    def "test interrupting its own thread does not kill test execution"() {
+        given:
+        buildFile << """
+            apply plugin: 'java'
+            ${mavenCentralRepository()}
+            dependencies {
+                ${testFrameworkDependencies}
+            }
+            test.${configureTestFramework}
+        """.stripIndent()
+
+        and:
+        file("src/test/java/SomeTest.java") << """
+            ${testFrameworkImports}
+
+            public class SomeTest {
+                @Test public void foo() {
+                    Thread.currentThread().interrupt();
+                }
+            }
+        """.stripIndent()
+
+        when:
+        run "test"
+
+        then:
+        executedAndNotSkipped(":test")
+    }
+
+    @Issue("https://issues.gradle.org/browse/GRADLE-2962")
+    def "incompatible user versions of classes that we also use don't affect test execution"() {
+
+        // These dependencies are quite particular.
+        // Both jars contain 'com.google.common.collect.ImmutableCollection'
+        // 'google-collections' contains '$EmptyImmutableCollection' which extends '$AbstractImmutableCollection' which is also in guava 15.
+        // In the google-collections version '$EmptyImmutableCollection' overrides `toArray()`.
+        // In guava 15, this method is final.
+        // This causes a verifier error when loading $EmptyImmutableCollection (can't override final method).
+
+        // Our test infrastructure loads org.gradle.util.SystemProperties, which depends on $EmptyImmutableCollection from guava 14.
+        // The below test is testing that out infrastructure doesn't throw a VerifyError while bootstrapping.
+        // This is testing classloader isolation, but this was not the real problem that triggered GRADLE-2962.
+        // The problem was that we tried to load the user's $EmptyImmutableCollection in a class loader structure we wouldn't have used anyway,
+        // but this caused the infrastructure to fail with an internal error because of the VerifyError.
+        // In a nutshell, this tests that we don't even try to load classes that are there, but that we shouldn't see.
+
+        when:
+        executer
+            .withArgument("-Porg.gradle.java.installations.paths=${AvailableJavaHomes.getAvailableJvms().collect { it.javaHome.absolutePath }.join(",")}")
+            .withToolchainDetectionEnabled()
+        buildFile << """
+            plugins {
+                id("java")
+            }
+            ${javaPluginToolchainVersion(11)}
+            ${mavenCentralRepository()}
+            configurations { first {}; last {} }
+            dependencies {
+                // guarantee ordering
+                first 'com.google.guava:guava:15.0'
+                last 'com.google.collections:google-collections:1.0'
+                implementation configurations.first + configurations.last
+
+                ${testFrameworkDependencies}
+            }
+            test.${configureTestFramework}
+        """.stripIndent()
+
+        and:
+        file("src/test/java/TestCase.java") << """
+            ${testFrameworkImports}
+            public class TestCase {
+                @Test
+                public void test() throws Exception {
+                    getClass().getClassLoader().loadClass("com.google.common.collect.ImmutableCollection\$EmptyImmutableCollection");
+                }
+            }
+        """.stripIndent()
+
+        then:
+        fails "test"
+
+        and:
+        def result = new DefaultTestExecutionResult(testDirectory)
+        result.testClass("TestCase").with {
+            assertTestFailed("test", CoreMatchers.containsString("java.lang.VerifyError"))
+            assertTestFailed("test", CoreMatchers.containsString("\$EmptyImmutableCollection"))
+        }
+    }
+
+    def "tests are re-executed when set of candidate classes change"() {
+        given:
+        buildFile << """
+            apply plugin:'java'
+            ${mavenCentralRepository()}
+            dependencies {
+                ${testFrameworkDependencies}
+            }
+            test {
+                ${configureTestFramework}
+                testLogging {
+                    events "passed", "skipped", "failed"
+                }
+            }
+        """.stripIndent()
+
+        and:
+        file("src/test/java/FirstTest.java") << """
+            ${testFrameworkImports}
+            public class FirstTest {
+                @Test public void test() {}
+            }
+        """.stripIndent()
+
+        file("src/test/java/SecondTest.java") << """
+            ${testFrameworkImports}
+            public class SecondTest {
+                @Test public void test() {}
+            }
+        """.stripIndent()
+
+        when:
+        run "test"
+        then:
+        executedAndNotSkipped ":test"
+        output.contains("FirstTest > ${maybeParentheses('test')} PASSED")
+        output.contains("SecondTest > ${maybeParentheses('test')} PASSED")
+
+        when:
+        run "test"
+        then:
+        skipped ":test"
+
+        when:
+        buildFile << """
+        test {
+            filter {
+                includeTestsMatching "First*"
+            }
+        }
+        """
+        then:
+        run "test"
+        then:
+        executedAndNotSkipped ":test"
+        output.contains("FirstTest > ${maybeParentheses('test')} PASSED")
+        !output.contains("SecondTest > ${maybeParentheses('test')} PASSED")
+    }
+
+    @Issue("https://github.com/gradle/gradle/issues/5305")
+    def "test can install an irreplaceable SecurityManager"() {
+        given:
+        executer
+            .withStackTraceChecksDisabled()
+            .withToolchainDetectionEnabled()
+        withInstallations(AvailableJavaHomes.getAvailableJvms())
+        buildFile << """
+            plugins {
+                id("java")
+            }
+            ${javaPluginToolchainVersion(11)}
+            ${mavenCentralRepository()}
+            dependencies {
+                ${testFrameworkDependencies}
+            }
+            test.${configureTestFramework}
+        """.stripIndent()
+
+        and:
+        file('src/test/java/SecurityManagerInstallationTest.java') << """
+            ${testFrameworkImports}
+            import java.security.Permission;
+
+            public class SecurityManagerInstallationTest {
+                @Test
+                public void testSecurityManagerCleanExit() {
+                    System.setSecurityManager(new SecurityManager() {
+                        @Override
+                        public void checkPermission(Permission perm) {
+                            if ("setSecurityManager".equals(perm.getName())) {
+                                throw new SecurityException("You cannot replace this security manager!");
+                            }
+                        }
+                    });
+                }
+            }
+        """.stripIndent()
+
+        when:
+        succeeds "test"
+
+        then:
+        outputContains "Unable to reset SecurityManager"
+    }
+
 }
